Allow fallback to hdpi assets when mdpi is missing
Google has deprecated MDPI in AOSP as no Nexus devices
have that density any more.  As a result, there are
an increasing number of missing assets in the tree
for this density.  This causes SystemUI to crash.

This allows HDPI assets to be used as a fallback
for MDPI assets that are not present.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>
Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Ie277c825ec6616b92892d4b500db887c8e6d8c31